requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
python-nmap==0.7.1
requests==2.32.4
urllib3==2.5.0
